+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Next
+
+- **[Fix]** Use `{}` instead of `Object.create(null)` when creating new objects in `DocumentType`.
+  Quickfix for [chaijs/deep-eql#51](https://github.com/chaijs/deep-eql/issues/51).
+
 # 0.6.0 (2018-01-29)
 
 - **[Breaking change]** Remove index module, you need to import specific modules with deep imports